# labels: improve Has() method for stringlabels build 